Test runtime with different type
I make a test for nested if and switch case to know who's faster.
Tested 3 times with same time interval, which is 5 seconds.

*on microsecond

Nested if	Switch case
604		459
274		451
504		295

If we do more test, it always giving a different results with almost the same difference.
Then the conclusion, switch case is more simpler than nested if.

Signed-off-by: Alwan Adiuntoro <alwanadiuntoro@gmail.com>